Add the structure and correct linters
