--- src/completed.js
@@ -1,13 +1,0 @@
-const checkBox = (index, items) => {
-  items[index].completed = !items[index].completed;
-  localStorage.localData = JSON.stringify(items);
-};
-
-export default (items) => {
-  const checkboxes = document.querySelectorAll('.checkbox');
-  checkboxes.forEach((checkbox, index) => {
-    checkbox.addEventListener('change', () => {
-      checkBox(index, items);
-    });
-  });
-};
