--- src/editDescription.js
@@ -1,9 +1,0 @@
-export default (items) => {
-  const taskDescriptions = document.querySelectorAll('.taskText');
-  taskDescriptions.forEach((task, index) => {
-    task.addEventListener('input', () => {
-      items[index].description = task.value;
-      localStorage.localData = JSON.stringify(items);
-    });
-  });
-};
