+++ src/index.js
@@ -1,16 +1,40 @@
 import './style.css';
-import { addTask, createList } from './addtask.js';
-import clearAll from './clearCompleted.js';
 
-let tasks = [];
+const items = [
+  {
+    description: 'Eat hamburgers',
+    completed: false,
+    index: 1,
+  },
+  {
+    description: 'Steal neighbours dog',
+    completed: false,
+    index: 3,
+  },
+  {
+    description: 'Become millionaire',
+    completed: false,
+    index: 2,
+  },
+];
 
-const getLocalData = () => {
-  if (localStorage.getItem('localData')) {
-    tasks = JSON.parse(localStorage.localData);
+const compare = (a, b) => {
+  if (a.index < b.index) {
+    return -1;
   }
+  if (a.index > b.index) {
+    return 1;
+  }
+  return 0;
 };
 
-getLocalData();
-createList(tasks);
-addTask(tasks);
-clearAll(tasks);
+const list = document.getElementById('todolist');
+
+const createList = () => {
+  items.sort(compare);
+  items.forEach((item) => {
+    list.innerHTML += `<div class="item"><input type="checkbox"><p>${item.description}</p><i class="fas fa-ellipsis-v"></i></div>`;
+  });
+};
+
+createList();
