+++ src/style.css
@@ -152,8 +152,7 @@ table {
   padding: 1em;
 }
 
-.taskText {
-  color: gray;
+p {
   margin: 0 0 0 1em;
 }
